learn api working

+++ SlitherModel/Source/Food.cs
@@ -13,7 +13,7 @@ namespace SlitherModel.Source
     public partial class Food
     {
         [JsonProperty("id")]
-        public long Id { get; set; }
+        public double Id { get; set; }
 
         [JsonProperty("xx")]
         public double Xx { get; set; }
@@ -28,46 +28,46 @@ namespace SlitherModel.Source
         public double Ry { get; set; }
 
         [JsonProperty("rsp")]
-        public long Rsp { get; set; }
+        public double Rsp { get; set; }
 
         [JsonProperty("cv")]
-        public long Cv { get; set; }
+        public double Cv { get; set; }
 
         [JsonProperty("rad")]
-        public long Rad { get; set; }
+        public double Rad { get; set; }
 
         [JsonProperty("sz")]
         public double Sz { get; set; }
 
         [JsonProperty("lrrad")]
-        public long Lrrad { get; set; }
+        public double Lrrad { get; set; }
 
         [JsonProperty("cv2")]
-        public long Cv2 { get; set; }
+        public double Cv2 { get; set; }
 
         [JsonProperty("fi")]
         public Fi Fi { get; set; }
 
         [JsonProperty("fw")]
-        public long Fw { get; set; }
+        public double Fw { get; set; }
 
         [JsonProperty("fh")]
-        public long Fh { get; set; }
+        public double Fh { get; set; }
 
         [JsonProperty("fw2")]
-        public long Fw2 { get; set; }
+        public double Fw2 { get; set; }
 
         [JsonProperty("fh2")]
-        public long Fh2 { get; set; }
+        public double Fh2 { get; set; }
 
         [JsonProperty("ofi")]
         public Fi Ofi { get; set; }
 
         [JsonProperty("ofw")]
-        public long Ofw { get; set; }
+        public double Ofw { get; set; }
 
         [JsonProperty("ofh")]
-        public long Ofh { get; set; }
+        public double Ofh { get; set; }
 
         [JsonProperty("ofw2")]
         public double Ofw2 { get; set; }
@@ -76,16 +76,16 @@ namespace SlitherModel.Source
         public double Ofh2 { get; set; }
 
         [JsonProperty("gcv")]
-        public long Gcv { get; set; }
+        public double Gcv { get; set; }
 
         [JsonProperty("gfi")]
         public Fi Gfi { get; set; }
 
         [JsonProperty("gfw")]
-        public long Gfw { get; set; }
+        public double Gfw { get; set; }
 
         [JsonProperty("gfh")]
-        public long Gfh { get; set; }
+        public double Gfh { get; set; }
 
         [JsonProperty("gfw2")]
         public double Gfw2 { get; set; }
@@ -94,16 +94,16 @@ namespace SlitherModel.Source
         public double Gfh2 { get; set; }
 
         [JsonProperty("g2cv")]
-        public long G2Cv { get; set; }
+        public double G2Cv { get; set; }
 
         [JsonProperty("g2fi")]
         public Fi G2Fi { get; set; }
 
         [JsonProperty("g2fw")]
-        public long G2Fw { get; set; }
+        public double G2Fw { get; set; }
 
         [JsonProperty("g2fh")]
-        public long G2Fh { get; set; }
+        public double G2Fh { get; set; }
 
         [JsonProperty("g2fw2")]
         public double G2Fw2 { get; set; }
@@ -112,7 +112,7 @@ namespace SlitherModel.Source
         public double G2Fh2 { get; set; }
 
         [JsonProperty("fr")]
-        public long Fr { get; set; }
+        public double Fr { get; set; }
 
         [JsonProperty("gfr")]
         public double Gfr { get; set; }
@@ -124,13 +124,13 @@ namespace SlitherModel.Source
         public double Wsp { get; set; }
 
         [JsonProperty("eaten_fr")]
-        public long EatenFr { get; set; }
+        public double EatenFr { get; set; }
 
         [JsonProperty("sx")]
-        public long Sx { get; set; }
+        public double Sx { get; set; }
 
         [JsonProperty("sy")]
-        public long Sy { get; set; }
+        public double Sy { get; set; }
 
         [JsonProperty("distance")]
         public double Distance { get; set; }
